gateway/requirements.txt


fastapi
uvicorn
requests
sqlalchemy
pydantic
python-multipart
python-dotenv
psycopg2-binary
